completed the css and html for the top bar of the purchase-box
