+++ client/src/index.jsx
@@ -1,25 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
+import NameTopBox from './components/topbaropen.jsx'
 
-
-class NameTopBox extends React.Component {
-    constructor(props) {
-        super(props)
-    }
-
-    render() {
-        return (
-            <div id='name-top-box'>
-                <div id='text-line'>
-                    <p className='buy'>Buy {this.props.ticker}</p>
-                    <p className='sell'>Sell {this.props.ticker}</p>
-                </div>
-            </div>
-        )   
-    }
-}
-
-
-ReactDOM.render(<NameTopBox ticker='SIRI'/>, document.getElementById('purchase-box'))
+ReactDOM.render(<NameTopBox ticker='AAPL'/>, document.getElementById('purchase-box'))
