+++ package.json
@@ -21,6 +21,7 @@
   },
   "homepage": "https://github.com/The-Merry-Men/service-purchase#readme",
   "dependencies": {
+    "eslint": "^5.8.0",
     "express": "^4.16.4",
     "faker": "^4.1.0",
     "mysql": "^2.16.0",
